Split symbol tree and symbol type tree into their own files
