+++ Cargo.toml
@@ -11,6 +11,7 @@ anyhow = "1.0.75"
 bytemuck = "1.16.1"
 chrono = "0.4.31"
 crossbeam-channel = "0.5.8"
+extended = "0.1.0"
 indexmap = "2.13.0"
 self_cell = "1.0.3"
 serde_json = "1.0.149"
